update upload files

+++ back-skeleton/src/main/java/com/pascal/backskeleton/controllers/PlaceController.java
@@ -1,10 +1,13 @@
 package com.pascal.backskeleton.controllers;
 
+import java.io.IOException;
+import java.sql.Timestamp;
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -14,11 +17,20 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
 
 import com.pascal.backskeleton.dao.PlaceRepository;
 import com.pascal.backskeleton.models.Place;
+import com.pascal.backskeleton.models.Place;
+import com.pascal.backskeleton.models.Place;
 
+import java.util.stream.Collectors;
+
+import com.pascal.backskeleton.services.FileStorageService;
+
+import io.micrometer.common.util.StringUtils;
 
 @RestController
 @RequestMapping("/api/places")
@@ -26,12 +38,22 @@ public class PlaceController {
 
     @Autowired
     private PlaceRepository placeRepository;
+    @Autowired
+    private FileStorageService fileStorageService;
 
     // Endpoint pour récupérer toutes les places
     @GetMapping
-    public ResponseEntity<List<Place>> getAllPlaces() {
+    public List<Place> getAllPlaces() {
+        // Récupérer tous les lieux
         List<Place> places = placeRepository.findAll();
-        return new ResponseEntity<>(places, HttpStatus.OK);
+
+        // Trier les lieux par ordre décroissant des ID
+        List<Place> sortedPlaces = places.stream()
+                .sorted((p1, p2) -> Long.compare(p2.getId(), p1.getId())) // Tri décroissant des ID
+                .collect(Collectors.toList());
+
+        // Retourner les lieux triés
+        return sortedPlaces;
     }
 
     // Endpoint pour récupérer une place par son ID
@@ -43,37 +65,92 @@ public class PlaceController {
     }
 
     // Endpoint pour créer une nouvelle place
-    @PostMapping
-    public ResponseEntity<Place> createPlace(@RequestBody Place place) {
+    @PostMapping(consumes = { MediaType.APPLICATION_OCTET_STREAM_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE })
+    public ResponseEntity<Place> createPlace(@RequestPart(name = "place") Place place,
+            @RequestPart(name = "imageUrl", required = false) MultipartFile file) {
+        place.setCreatedAt(new Timestamp(System.currentTimeMillis()));
+        place.setUpdatedAt(new Timestamp(System.currentTimeMillis()));
+
+        // Si un fichier est fourni, associez-le à l'utilisateur
+        if (file != null && !file.isEmpty()) {
+            try {
+                // Sauvegardez le fichier et récupérez son URL
+                String fileUrl = fileStorageService.storeFile(file, "places/");
+
+                // Associez l'URL du fichier à l'utilisateur
+
+                place.setImageUrl("/assets/images/places/" + fileUrl);
+
+            } catch (IOException e) {
+                e.printStackTrace();
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+            }
+        }
+
         Place savedPlace = placeRepository.save(place);
         return new ResponseEntity<>(savedPlace, HttpStatus.CREATED);
     }
 
-    // Endpoint pour mettre à jour une place existante
-    @PutMapping("/{id}")
-    public ResponseEntity<Place> updatePlace(@PathVariable("id") Long id, @RequestBody Place place) {
+    @PutMapping(value = "/{id}", consumes = { MediaType.MULTIPART_FORM_DATA_VALUE, MediaType.APPLICATION_OCTET_STREAM_VALUE })
+    public ResponseEntity<Place> updatePlace(@PathVariable Long id, @RequestPart("place") Place newPlace,
+            @RequestPart(value = "imageUrl", required = false) MultipartFile file) {
+    
         Optional<Place> optionalPlace = placeRepository.findById(id);
+    
         if (optionalPlace.isPresent()) {
             Place existingPlace = optionalPlace.get();
-            existingPlace.setTitle(place.getTitle());
-            existingPlace.setAddress(place.getAddress());
-            existingPlace.setImageUrl(place.getImageUrl());
-            existingPlace.setOpeningHours(place.getOpeningHours());
-            existingPlace.setCreatedAt(place.getCreatedAt());
-
-            Place updatedPlace = placeRepository.save(existingPlace);
-            return new ResponseEntity<>(updatedPlace, HttpStatus.OK);
+            existingPlace.setTitle(newPlace.getTitle());
+            existingPlace.setAddress(newPlace.getAddress());
+            existingPlace.setOpeningHours(newPlace.getOpeningHours());
+            existingPlace.setUpdatedAt(newPlace.getUpdatedAt());
+    
+            try {
+                String fileUrl = null;
+    
+                if (file != null && !file.isEmpty()) {
+                    // Sauvegarder le fichier et récupérer son URL
+                    if (StringUtils.isNotBlank(existingPlace.getImageUrl())) {
+                        fileUrl = fileStorageService.updateFile(existingPlace.getImageUrl(), file, "places/");
+                    } else {
+                        fileUrl = fileStorageService.storeFile(file, "places/");
+                    }
+                    
+                    // Associer l'URL du fichier à l'utilisateur
+                    existingPlace.setImageUrl("/assets/images/places/" + fileUrl);
+                }
+    
+                // Enregistrer l'utilisateur mis à jour dans la base de données
+                Place updatedPlace = placeRepository.save(existingPlace);
+                return new ResponseEntity<>(updatedPlace, HttpStatus.OK);
+            } catch (IOException e) {
+                e.printStackTrace();
+                return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+            }
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
-
     // Endpoint pour supprimer une place
     @DeleteMapping("/{id}")
     public ResponseEntity<HttpStatus> deletePlace(@PathVariable("id") Long id) {
         try {
-            placeRepository.deleteById(id);
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            Optional<Place> optionalPlace = placeRepository.findById(id);
+
+            if (optionalPlace.isPresent()) {
+                Place place = optionalPlace.get();
+
+                // Supprimer l'image du lieu s'il en a une
+                if (StringUtils.isNotBlank(place.getImageUrl())) {
+                    fileStorageService.deleteFile(place.getImageUrl());
+                }
+
+                // Supprimer le lieu de la base de données
+                placeRepository.deleteById(id);
+
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
